refactor: purge MobileLayout + ResponsiveLayout; promote single LandingPage
Collapse the dual desktop/mobile UI into one layout. The app previously
branched on window width in ResponsiveLayout, maintaining two parallel
component trees; this removes that split so there is a single layout to
maintain and (in later phases) make responsive across all screen sizes.

- Delete src/components/layout/MobileLayout (6 subcomponents + 2.2k-line
  CSS) and ResponsiveLayout (the width-based device switch).
- Rename DesktopLayout -> LandingPage (folder, files, and component
  HomeScreen -> LandingPage): it's the single view a visitor sees at
  demaceo.com regardless of device.
- Point src/app/page.tsx straight at LandingPage; update layout barrel.
- Remove now-dead html.touch-performance-mode rules from ModalFrame
  (the class was only ever set by MobileLayout) and stale commented
  device-hiding media queries.
- Update README architecture references.

No behavior change for desktop/tablet; small screens now receive the
same layout (responsive refinement tracked as a follow-up phase).
Home route bundle drops ~15kB (54.3 -> 39.1 kB First Load JS).
