+++ src/app/page.tsx
@@ -1,6 +1,6 @@
-import { ResponsiveLayout } from "@/components/layout";
+import { LandingPage } from "@/components/layout";
 import "./globals.css";
 
 export default function Home() {
-  return <ResponsiveLayout />;
+  return <LandingPage />;
 }
