+++ src/components/features/resume/InteractiveResume/InteractiveResume.css
@@ -957,7 +957,7 @@
 }
 
 /* Dark Mode Support */
-/* Dark mode override removed — DesktopLayout is always light-themed */
+/* Dark mode override removed — LandingPage is always light-themed */
 
 /* Touch-specific optimizations */
 @media (hover: none) and (pointer: coarse) {
