+++ src/components/layout/LandingPage/LandingPage.css
@@ -412,14 +412,3 @@
 }
 
 /* Responsive adjustments */
-/* @media (max-width: 768px) {
-  .macintosh-container {
-    display: none;
-  }
-} */
-
-/* @media (min-width: 769px) {
-  .iphone-container {
-    display: none;
-  }
-} */
