+++ src/components/layout/LandingPage/components/DesktopIcons.tsx
