--- src/components/layout/MobileLayout/components/PageIndicators.tsx
@@ -1,14 +1,0 @@
-"use client";
-
-import React from "react";
-
-const PageIndicators: React.FC = () => {
-  return (
-    <div className="page-indicators">
-      <div className="page-dot active"></div>
-      <div className="page-dot"></div>
-    </div>
-  );
-};
-
-export default PageIndicators;
