--- src/components/layout/MobileLayout/components/index.ts
@@ -1,6 +1,0 @@
-export { default as StatusBar } from './StatusBar';
-export { default as MobileWelcomeWindow } from './MobileWelcomeWindow';
-export { default as HomeApps } from './HomeApps';
-export { default as PageIndicators } from './PageIndicators';
-export { default as MobileDock } from './MobileDock';
-export { default as MobileModals } from './MobileModals';
